mmf/requirements.txt
numpy
tqdm>=4.43.0,<4.50.0
GitPython==3.1.0
requests==2.23.0
fasttext==0.9.1
nltk==3.4.5
editdistance==0.5.3
transformers>=3.4.0, <=4.10.1
sklearn==0.0
omegaconf>=2.0.6, <=2.1
lmdb==0.98
termcolor==1.1.0
iopath==0.1.8
datasets==1.2.1
matplotlib==3.3.4
pycocotools==2.0.4
ftfy==5.8
pytorch-lightning @ git+https://github.com/PyTorchLightning/pytorch-lightning@9b011606f
pillow==8.3.1
sentencepiece==0.1.86
opencv-python
networkx
gensim
psutil
filelock
six